refactor(_tools): extract panel_taxonomy.py from quality_scorer.py
Behavior-preserving refactor for epic #1625 (Content Coverage Audit).
Moves shared panel-taxonomy constants and scholar-key discovery into
a new module so the upcoming coverage_auditor.py (#1628) can reuse
them without duplication.

Additions beyond the moves (for #1628 consumption):
- LINGUISTIC_PANELS / HISTORICAL_PANELS / CONNECTIVE_PANELS
- GENRE_EXPECTED_PANELS (evidence-based per-genre rubric)
- TEMPLATE_PATTERNS (populated by A3 curation pass)
- categorize_panel() helper

accuracy_config.py is intentionally untouched — its taxonomy is a
deliberate separate surface for the accuracy auditor.

Byte-identical output validation passed:
  diff qs_stdout_before qs_stdout_after  -> (identical)
  diff qs_stderr_before qs_stderr_after  -> (identical)
  diff qs_json_before   qs_json_after    -> (identical)

Closes #1627.
